rl-manager/grpc_cloudsimplus/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "grpc_cloudsimplus"
version = "0.1.0"
description = "CloudSim Plus gRPC bridge for Stable Baselines3"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
]
